Reproduce the missing line issue
