--- src/test/java/tdl/record/sourcecode/ConvertToGitEdgeCasesTest.java
@@ -1,179 +1,0 @@
-package tdl.record.sourcecode;
-
-import org.eclipse.jgit.api.Git;
-import org.junit.Assert;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
-import support.TestGeneratedSrcsFile;
-import tdl.record.sourcecode.snapshot.SnapshotTypeHint;
-import tdl.record.sourcecode.snapshot.helpers.GitHelper;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Arrays;
-import java.util.Collections;
-
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasItems;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
-import static support.TestUtils.writeFile;
-
-public class ConvertToGitEdgeCasesTest {
-
-    @Rule
-    public TemporaryFolder folder = new TemporaryFolder();
-
-    @Rule
-    public TestGeneratedSrcsFile srcsFileWithRename = new TestGeneratedSrcsFile(Arrays.asList(
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "MSG1");
-                return SnapshotTypeHint.KEY;
-            },
-            // Case 1 = Patch with rename
-            (Path dst) -> {
-                writeFile(dst, "Test.txt", "MSG2");
-                return SnapshotTypeHint.PATCH;
-            }
-    ), Collections.emptyList());
-
-    @Rule
-    public TestGeneratedSrcsFile srcsFileWithFinalLineRemoval = new TestGeneratedSrcsFile(Arrays.asList(
-            // Step 1 - content with line plus ending statement
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY\n\ncontent-no-newline");
-                return SnapshotTypeHint.KEY;
-            },
-            // Step 1 - remove the ending statement
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY\n\n");
-                return SnapshotTypeHint.PATCH;
-            },
-            // Step 2 - re-add the ending statement
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY\n\ncontent-with-newline");
-                return SnapshotTypeHint.PATCH;
-            }), Collections.emptyList());
-
-
-    @Rule
-    public TestGeneratedSrcsFile srcsFileWithMissingLineRemoval = new TestGeneratedSrcsFile(Arrays.asList(
-            // Step 1 - content with no line ending
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY");
-                return SnapshotTypeHint.KEY;
-            }, //key
-            // Step 2 - content with 3 new lines and then a line of text
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY\n\n\ncheckout");
-                return SnapshotTypeHint.PATCH;
-            }, //patch
-            // Step 3 - content with 1 new line removed from the middle
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY\n\ncheckout");
-                return SnapshotTypeHint.PATCH;
-            }, //patch
-            // Step 4 - content with last two lines removed (new line and a line of text)
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY\n");
-                return SnapshotTypeHint.PATCH;
-            }, //patch
-            // Step 5 - content with new line removed from the end (back to original file)
-            (Path dst) -> {
-                writeFile(dst, "test.txt", "BODY");
-                return SnapshotTypeHint.PATCH;
-            } //patch
-    ), Collections.emptyList());
-
-    @Test
-    public void patchWithRenameShouldIgnoreTheRename() throws Exception {
-        ConvertToGitCommand command = new ConvertToGitCommand();
-        File outputDir = folder.newFolder();
-
-        command.inputFilePath = srcsFileWithRename.getFilePath().toString();
-        command.outputDirectoryPath = outputDir.toString();
-        command.run();
-
-        Git git = Git.init().setDirectory(outputDir).call();
-        assertThat("Does not have the commits",
-                GitHelper.getCommitCount(git), equalTo(2));
-
-        String commit = "HEAD";
-        git.checkout().setName("master").setStartPoint(commit).call();
-        Path testFile = outputDir.toPath().resolve("Test.txt");
-        assertTrue("File  " + testFile + " does not exist in " + commit,
-                testFile.toFile().exists());
-
-        assertThat(Files.readAllLines(testFile), hasItems("MSG2"));
-    }
-
-    @Test
-    public void usersShouldBeAbleToSafelyRemoveTheLastLine() throws Exception {
-        ConvertToGitCommand command = new ConvertToGitCommand();
-        File outputDir = folder.newFolder();
-
-        command.inputFilePath = srcsFileWithFinalLineRemoval.getFilePath().toString();
-        command.outputDirectoryPath = outputDir.toString();
-        command.run();
-    }
-
-    @Test
-    public void usersShouldBeAbleToSaveFilesWithNoLineEndings() throws Exception {
-        ConvertToGitCommand command = new ConvertToGitCommand();
-        File outputDir = folder.newFolder();
-
-        command.inputFilePath = srcsFileWithMissingLineRemoval.getFilePath().toString();
-        command.outputDirectoryPath = outputDir.toString();
-        command.run();
-    }
-
-    @Test
-    public void applyPatchToSourceWhenTheLastLineIsMissingFromTheHunk() throws Exception {
-//  Export segment 7222L from the sourcecode_20180405T111942.srcs
-//        ExportSegmentsCommand exportSegmentsCommand = new ExportSegmentsCommand(
-//                "/path/to/sourcecode_20180405T111942.srcs",
-//                "/path/to/dev-sourcecode-record/src/test/resources/array_index_out_of_bounds_folder/git_issue_24/snapshots.srcs",
-//                Arrays.asList(7402L));
-//        exportSegmentsCommand.run();
-        replayApplyingCommitUsingSrcsSegementsFor("git_issue_24");
-    }
-
-    @Test
-    public void applyPatchToSourceWhenTheLastLineIsMissingBeforeHunkCanBeApplied() throws Exception {
-//  Export segment 3224L from the sourcecode_20180811T125110.srcs
-//        ExportSegmentsCommand exportSegmentsCommand = new ExportSegmentsCommand(
-//                "/path/to/sourcecode_20180811T125110.srcs",
-//                "/path/to/dev-sourcecode-record/src/test/resources/array_index_out_of_bounds_folder/git_issue_25/snapshots.srcs",
-//                Arrays.asList(3224L));
-//        exportSegmentsCommand.run();
-        replayApplyingCommitUsingSrcsSegementsFor("git_issue_25");
-    }
-
-    private void replayApplyingCommitUsingSrcsSegementsFor(String gitIssue) throws IOException {
-        ListCommand listCommand = new ListCommand();
-        String srcsFilePath = getSrcsFileFor(gitIssue);
-        listCommand.inputFilePath = srcsFilePath;
-        listCommand.run();
-
-        ConvertToGitCommand command = new ConvertToGitCommand();
-        File outputDir = folder.newFolder();
-
-        command.inputFilePath = srcsFilePath;
-        command.outputDirectoryPath = outputDir.toString();
-
-        try {
-            command.run();
-        } catch (Exception ex) {
-            Assert.fail("Did not complete commit, due to exception: " + ex.getMessage());
-        }
-    }
-
-    private String getSrcsFileFor(final String gitIssue) {
-        ClassLoader classLoader = ConvertToGitEdgeCasesTest.class.getClassLoader();
-        File file = new File(classLoader.getResource("array_index_out_of_bounds_folder/" + gitIssue + "/snapshots.srcs").getFile());
-        return file.toPath().toString();
-    }
-}
